+++ src/MixMaster.cpp
@@ -610,12 +610,14 @@ struct MixMaster : Module {
 		// Aux sends (send track and group audio (16+4 or 8+2 stereo signals) to auxspander
 		// auxSends[] has room for 16+4 or 8+2 stereo values of the sends to the aux panel (Trk1L, Trk1R, Trk2L, Trk2R ... Trk16L, Trk16R, Grp1L, Grp1R ... Grp4L, Grp4R)
 		// populate auxSends[0..39]: Take the trackTaps/groupTaps indicated by the Aux sends mode (with per-track option)
+				
+		float masterGain = gInfo.masterFaderScalesSends == 0 ? 1.0f : master.gainMatrixSlewers.out[1] + master.gainMatrixSlewers.out[2];
 		
 		// tracks
-		if ( gInfo.auxSendsMode < 4 && (gInfo.groupsControlTrackSendLevels == 0 || gInfo.groupUsage[N_GRP] == 0) ) {
+		if ( gInfo.auxSendsMode < 4 && gInfo.masterFaderScalesSends == 0 && (gInfo.groupsControlTrackSendLevels == 0 || gInfo.groupUsage[N_GRP] == 0) ) {
 			memcpy(auxSends, &trackTaps[gInfo.auxSendsMode << (3 + N_TRK / 8)], (N_TRK * 2) * 4);
 		}
-		else {
+		else {			
 			int trkGroup;
 			for (int trk = 0; trk < N_TRK; trk++) {
 				// tracks should have aux sends even when grouped
@@ -630,21 +632,25 @@ struct MixMaster : Module {
 					trackR = sigs[1] + sigs[3];
 					trackL *= groups[trkGroup].muteSoloGainSlewer.out;
 					trackR *= groups[trkGroup].muteSoloGainSlewer.out;
-				}					
+				}			
+				if (masterGain != 1.0f) {
+					trackL *= masterGain;
+					trackR *= masterGain;
+				}
 				auxSends[(trk << 1) + 0] = trackL;
 				auxSends[(trk << 1) + 1] = trackR;
 			}
 		}
 		
 		// groups
-		if (gInfo.auxSendsMode < 4) {
+		if (gInfo.auxSendsMode < 4 && gInfo.masterFaderScalesSends == 0) {
 			memcpy(&auxSends[N_TRK * 2], &groupTaps[gInfo.auxSendsMode << (1 + N_GRP / 2)], (N_GRP * 2) * 4);
 		}
 		else {
 			for (int grp = 0; grp < N_GRP; grp++) {
 				int tapIndex = (groups[grp].auxSendsMode << (1 + N_GRP / 2));
-				auxSends[(grp << 1) + N_TRK * 2] = groupTaps[tapIndex + (grp << 1) + 0];
-				auxSends[(grp << 1) + N_TRK * 2] = groupTaps[tapIndex + (grp << 1) + 1];
+				auxSends[(grp << 1) + N_TRK * 2] = groupTaps[tapIndex + (grp << 1) + 0] * masterGain;
+				auxSends[(grp << 1) + N_TRK * 2] = groupTaps[tapIndex + (grp << 1) + 1] * masterGain;
 			}
 		}
 
@@ -1109,6 +1115,7 @@ struct MixMasterWidget : ModuleWidget {
 			masterDisplay->dimGainIntegerDB = &(module->master.dimGainIntegerDB);
 			masterDisplay->colorAndCloak = &(module->gInfo.colorAndCloak);
 			masterDisplay->idSrc = &(module->id);
+			masterDisplay->masterFaderScalesSendsSrc = &(module->gInfo.masterFaderScalesSends);
 		}
 		
 		// Master fader
@@ -1396,6 +1403,7 @@ struct MixMasterJrWidget : ModuleWidget {
 			masterDisplay->dimGainIntegerDB = &(module->master.dimGainIntegerDB);
 			masterDisplay->colorAndCloak = &(module->gInfo.colorAndCloak);
 			masterDisplay->idSrc = &(module->id);
+			masterDisplay->masterFaderScalesSendsSrc = &(module->gInfo.masterFaderScalesSends);
 		}
 		
 		// Master fader