+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added initialization of individual track settings in MixMasters and EqMaster
 - Made filter mappable (click filter lights when mapping)
 - Added option in EQMaster menu to hide EQ curves when bypassed, for when EQ is to be used as a spectrum analyzer only
+- Added option in MixMaster to make master fader scale aux sends (for chaining mixers with shared aux effects)
 
 
 ### 1.1.5 (2020-01-29)